we dont need to try and delete the site's root category if it's already been deleted
git-svn-id: http://svn.krangcms.com/trunk/krang@5939 689d98c0-9023-0410-9c0b-f79783857903

+++ lib/Krang/Site.pm
@@ -259,7 +259,7 @@ sub delete {
         dir     => '/',
         site_id => $id
     );
-    $root->delete();
+    $root->delete() if $root; # only need to do this if the category hasn't already been deleted
 
     # remove record from the site table
     my $dbh = dbh();
